% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_plot.R
\name{ci_plot}
\alias{ci_plot}
\title{Produce a confidence interval (CI) plot for predicted means}
\usage{
ci_plot(plot_mt, mod_df = NULL, resp_name = NULL, jitterv = 0.2, basesz = 12)
}
\arguments{
\item{plot_mt}{Data frame of the mean table which is the output of \code{mean_table} or \code{mean_table$Table} from function \code{predictmeans}.}

\item{mod_df}{Data frame of the data used in the modelling. The default is NULL, then plot will has CIs only.}

\item{resp_name}{Name (in "quotes") for the (original) response variable used in the modelling.}

\item{jitterv}{A degree of jitter in x and y direction in the back transformed means graph. The default is 0.2.}

\item{basesz}{Base font size, given in pts.}
}
\description{
Produce a confidence interval (CI) plot for predicted means
}
\examples{

library(predictmeans)
ftable(xtabs(yield ~ Block+Variety+nitro, data=Oats))
Oats$nitro <- factor(Oats$nitro)
fm <- lme(yield ~ nitro*Variety, random=~1|Block/Variety, data=Oats)
#--------------------------------------------------------
predm_out <- predictmeansN(fm, "nitro", adj="BH")
ci_plot(predm_out$mean_table$Table)
ci_plot(predm_out$mean_table$Table, mod_df = Oats, resp_name = "yield")
# Use plot method
plot(predm_out, mod_df = Oats, resp_name = "yield")
#--------------------------------------------------------
predictout <- predictmeansN(fm, "nitro:Variety", atvar="Variety", adj="BH")
plot(predictout, mod_df = Oats, resp_name = "yield", line=FALSE)

}
\author{
Dongwen Luo
}
