\name{Clinical}
\alias{Clinical}
\docType{data}
\title{\code{Clinical} data}

\description{\code{Clinical} data is from a multicentre randomized clinical trial (Beitler
 & Landis 1985, Biometrics).
}
\usage{
data(Clinical)
}
%- maybe also `usage' for other objects documented here.

\format{\code{Clinical} is a 16 row data frame with the following columns 
\describe{
\item{Clinic}{8 centres id.}
\item{Treatment}{2 skin treatments (control or active drug).}
\item{Favorable}{number that produced a favourable response.}
\item{Total}{number of patients in each treatment group.}
}
} 

