% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{model.frame.gls}
\alias{model.frame.gls}
\title{model.frame method for objects of class 'gls'}
\usage{
\method{model.frame}{gls}(formula, ...)
}
\arguments{
\item{formula}{An object of class \code{gls}.}

\item{...}{Further arguments passed to \code{model.frame.formula}.}
}
\description{
This is a specific \code{model.frame} method for objects of the class
\code{gls}. It simply calls \code{nlme::getData}. See \link[nlme]{getData}.
}
