% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{model.frame.lme}
\alias{model.frame.lme}
\title{model.frame method for objects of class 'lme'}
\usage{
\method{model.frame}{lme}(formula, ...)
}
\arguments{
\item{formula}{An object of class \code{lme}.}

\item{...}{Further arguments passed to \code{model.frame.formula}.}
}
\description{
This is a specific \code{model.frame} method for objects of the class \code{lme}.
See \link[nlme]{lme}.
}
