% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{model.matrix.aovlist}
\alias{model.matrix.aovlist}
\title{model.matrix method for objects of class 'aovlist'}
\usage{
\method{model.matrix}{aovlist}(object, ...)
}
\arguments{
\item{object}{An object of class \code{aovlist}.}

\item{...}{Not used.}
}
\description{
This is a specific \code{model.matrix} method for objects of the class
\code{aovlist}. It does not do anything except give an error because
\code{aovlist} is not supported.
}
