% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{model.matrix.gls}
\alias{model.matrix.gls}
\title{model.matrix method for objects of class 'gls'}
\usage{
\method{model.matrix}{gls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{gls}.}

\item{...}{Further arguments passed to \code{model.matrix.terms}.}
}
\description{
This is a specific \code{model.matrix} method for objects of the class
\code{gls}. See \link[nlme]{getData}.
}
