% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{model.matrix.lme}
\alias{model.matrix.lme}
\title{model.matrix method for objects of class 'lme'}
\usage{
\method{model.matrix}{lme}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lme}.}

\item{...}{Further arguments passed to \code{model.matrix.terms}.}
}
\description{
This is a specific \code{model.frame} method for objects of the class
\code{lme}. See \link[nlme]{lme}.
}
