% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pdmlist.R
\name{plot.pdmlist}
\alias{plot.pdmlist}
\title{Plot the predicted means}
\usage{
\method{plot}{pdmlist}(x, type = c("bar", "mean", "pm"), ...)
}
\arguments{
\item{x}{an object of class \code{pdmlist}. See \link{list("predictmeans")}
for more info.}

\item{type}{Plot type. Can be one of \code{"bar"}, \code{"mean"},
\code{"pm"}. One letter abbrevitions, i.e. \code{"b"}, \code{"m"},
\code{"pm"} will work as well.}

\item{...}{other arguments}
}
\description{
Plot the predicted means
}
