% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictmeans-package.R
\docType{_PACKAGE}
\name{predictmeans-package}
\alias{predictmeans-package}
\title{Predicted Means for Linear and Semiparametric Models}
\description{
This package provides functions to diagnose and make inferences from various
linear models, such as those obtained from 'aov', 'lm', 'glm', 'gls', 'lme',
'lmer', 'glmmTMB' and 'semireg'. Inferences include predicted means and
standard errors, contrasts, multiple comparisons, permutation tests,
adjusted R-square and graphs.
}
\details{
\tabular{ll}{ Package: \tab predictmeans\cr Type: \tab Package\cr Version:
\tab 1.1.1\cr Date: \tab 2024-04-03\cr License: \tab GPL (>= 2)\cr }
}
\references{
Welham, S., Cullis, B., Gogel, B., Gilmour, A., & Thompson, R.
(2004), \emph{Prediction in linear mixed models}, Australian and New Zealand
Journal of Statistics, 46(3), 325-347.
}
\author{
Dongwen Luo, Siva Ganesh and John Koolaard

Maintainer: Dongwen Luo <dongwen.luo@agresearch.co.nz>
}
\keyword{package}
