% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pdmlist.R
\name{print.pdmlist}
\alias{print.pdmlist}
\title{print method for objects of class 'pdmlist'}
\usage{
\method{print}{pdmlist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pdmlist}.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
This is a specific \code{print} method for objects of the class
\code{pdmlist}. See \code{\link{predictmeans}}.
}
