% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{terms.gls}
\alias{terms.gls}
\title{terms method for objects of class 'gls'}
\usage{
\method{terms}{gls}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gls}.}

\item{...}{Additional arguments passed to \code{terms}.}
}
\description{
This is a specific \code{terms} method for objects of the class \code{gls}.
See \link[nlme]{getData}.
}
