% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.parm.functs.R
\name{terms.lme}
\alias{terms.lme}
\title{terms method for objects of class 'lme'}
\usage{
\method{terms}{lme}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lme}.}

\item{...}{Not used.}
}
\description{
This is a specific \code{terms} method for objects of the class \code{lme}.
See \link[nlme]{lme}.
}
