% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictmeans-package.R
\docType{data}
\name{Clinical}
\alias{Clinical}
\title{\code{Clinical} data}
\format{
\code{Clinical} is a 16 row data frame with the following columns
\describe{ \item{Clinic}{8 centres id.} \item{Treatment}{2 skin treatments
(control or active drug).} \item{Favorable}{number that produced a
favourable response.} \item{Total}{number of patients in each treatment
group.} }
}
\description{
\code{Clinical} data is from a multicentre randomized clinical trial
(Beitler & Landis 1985, Biometrics).
}
