\name{predictmeans-package}
\alias{predictmeans-package}
\docType{package}
\title{
Calculate Predicted Means for Linear Models
}
\description{
This package provides functions to diagnose and make inferences from 
various linear models, such as those obtained from 'aov', 'lm', 'glm', 
'gls', 'lme', and 'lmer'. Inferences include predicted means and standard 
errors, contrasts, multiple comparisons, permutation tests and graphs. 
}
\details{
\tabular{ll}{
Package: \tab predictmeans\cr
Type: \tab Package\cr
Version: \tab 0.98\cr
Date: \tab 2014-03-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Dongwen Luo, Siva Ganesh and John Koolaard

Maintainer: Dongwen Luo <dongwen.luo@agresearch.co.nz>
}
\references{
	Welham, S., Cullis, B., Gogel, B., Gilmour, A., & Thompson, R. (2004),
	\emph{Prediction in linear mixed models}, Australian and New Zealand Journal of Statistics, 46(3), 325-347. 

}

\keyword{ package }
